MAINT: Change behaviors of int and bool in spacing

+++ numpy/core/code_generators/ufunc_docstrings.py
@@ -3416,7 +3416,7 @@ add_newdoc('numpy.core.umath', 'spacing',
 
     Examples
     --------
-    >>> np.spacing(1) == np.finfo(np.float64).eps
+    >>> np.spacing(np.float64(1)) == np.finfo(np.float64).eps
     True
 
     """)
